refactor(render-worker example): move logic lib.rs → worker.rs
lib.rs keeps only the crate attr + crate doc + mod worker; pub use worker::*.
All boot logic + WorkerInputEvent protocol moved to worker.rs verbatim. Zero
behavior change (cargo check + clippy clean).

Completes the mod.rs/lib.rs sweep: all 6 files that held logic (scene-schema,
frame_globals, pipeline_scheduler, dynamic_materials, renderer lib.rs,
render-worker) now keep their roots to module-exposure only.
